requirements.txt
asgiref==3.7.2
cfgv==3.4.0
distlib==0.3.7
Django==4.2.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.12.4
identify==2.5.30
nodeenv==1.8.0
platformdirs==3.11.0
pre-commit==3.5.0
psycopg2-binary==2.9.6
PyJWT==2.7.0
python-decouple==3.8
pytz==2023.3
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.6.3
virtualenv==20.24.5
